// src/Interfaces/Wine.ts
interface Wine {
  id: number;
  name: string | null;
  rating: number | null;
  rating_count: number | null;
  image_url: string | null;
  price: number | null;
  country: string | null;
  district: string | null;
  sub_district: string | null;
  type: string | null;
  url: string | null;
  vivino_url: string | null;
  volume: number | null;
  updated_at: string | null;
  count: number | null;
}

export default Wine;